requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
pandas==3.0.0
openpyxl==3.1.5
python-dotenv==1.2.1
openai>=1.0.0
google-generativeai>=0.3.0
gunicorn>=21.0.0
whitenoise>=6.6.0

